requirements.txt
pyserial>=3.4
numpy>=1.18.0
pytest>=5.3.5
